Use fail instead of error
`error` in Template Haskell splices should be used to report bugs
in the Template Haskell code, and ideally not to report bugs in
the user code calling it. When GHC encounters `error`, it gives a
rather more verbose (and ugly) error report, based on that assumption.
The usual way to report a problem in this context is to use `fail`.